+++ EleCfitsioWrapper/tests/src/ImageWrapper_test.cpp
@@ -38,7 +38,7 @@ void checkRandom3DRasterIsReadBack() {
   using namespace Fits::Test;
   RandomRaster<T, 3> input({ 2, 3, 4 });
   MinimalFile file;
-  HduAccess::createImageExtension(file.fptr, "IMGEXT", input);
+  HduAccess::assignImageExtension(file.fptr, "IMGEXT", input);
   const auto fixedOutput = ImageIo::readRaster<T, 3>(file.fptr);
   BOOST_TEST(fixedOutput.vector() == input.vector());
   const auto variableOuptut = ImageIo::readRaster<T, -1>(file.fptr);
@@ -62,7 +62,7 @@ BOOST_FIXTURE_TEST_CASE(region_is_read_back, Fits::Test::MinimalFile) {
       }
     }
   }
-  HduAccess::createImageExtension(fptr, "EXT", input);
+  HduAccess::assignImageExtension(fptr, "EXT", input);
   const auto region = Fits::Region<3>::fromShape({ 1, 0, 1 }, { 2, 3, 3 });
   const auto view = ImageIo::readRegion<long, 3>(fptr, region);
   BOOST_TEST(view.shape() == region.shape());